+++ section/section.go
@@ -8,13 +8,13 @@ import (
 )
 
 // SectionOpt customizes a Section on creation
-type SectionOpt func(*cli.SectionT)
+type SectionOpt func(*cli.Section)
 
 // NewSectionT creates a Section!
-func NewSectionT(helpShort string, opts ...SectionOpt) cli.SectionT {
-	section := cli.SectionT{
+func NewSectionT(helpShort string, opts ...SectionOpt) cli.Section {
+	section := cli.Section{
 		HelpShort: helpShort,
-		Sections:  make(cli.SectionMapT),
+		Sections:  make(cli.SectionMap),
 		Commands:  make(cli.CommandMap),
 		HelpLong:  "",
 		Footer:    "",
@@ -26,8 +26,8 @@ func NewSectionT(helpShort string, opts ...SectionOpt) cli.SectionT {
 }
 
 // Section adds an existing Section underneath this Section. Panics if a Section with the same name already exists
-func Section(name string, value cli.SectionT) SectionOpt {
-	return func(app *cli.SectionT) {
+func Section(name string, value cli.Section) SectionOpt {
+	return func(app *cli.Section) {
 		if _, alreadyThere := app.Sections[name]; !alreadyThere {
 			app.Sections[name] = value
 		} else {
@@ -37,8 +37,8 @@ func Section(name string, value cli.SectionT) SectionOpt {
 }
 
 // SectionMap adds existing Sections underneath this Section. Panics if a Section with the same name already exists
-func SectionMap(sections cli.SectionMapT) SectionOpt {
-	return func(app *cli.SectionT) {
+func SectionMap(sections cli.SectionMap) SectionOpt {
+	return func(app *cli.Section) {
 		for name, value := range sections {
 			Section(name, value)(app)
 		}
@@ -47,7 +47,7 @@ func SectionMap(sections cli.SectionMapT) SectionOpt {
 
 // Command adds an existing Command underneath this Section. Panics if a Command with the same name already exists
 func Command(name string, value cli.Command) SectionOpt {
-	return func(app *cli.SectionT) {
+	return func(app *cli.Section) {
 		if _, alreadyThere := app.Commands[name]; !alreadyThere {
 			app.Commands[name] = value
 		} else {
@@ -58,7 +58,7 @@ func Command(name string, value cli.Command) SectionOpt {
 
 // CommandMap adds existing Commands underneath this Section. Panics if a Command with the same name already exists
 func CommandMap(commands cli.CommandMap) SectionOpt {
-	return func(app *cli.SectionT) {
+	return func(app *cli.Section) {
 		for name, value := range commands {
 			Command(name, value)(app)
 		}
@@ -77,14 +77,14 @@ func NewCommand(name string, helpShort string, action cli.Action, opts ...comman
 
 // Footer adds an optional help string to this Section
 func Footer(footer string) SectionOpt {
-	return func(cat *cli.SectionT) {
+	return func(cat *cli.Section) {
 		cat.Footer = footer
 	}
 }
 
 // HelpLong adds an optional help string to this Section
 func HelpLong(helpLong string) SectionOpt {
-	return func(cat *cli.SectionT) {
+	return func(cat *cli.Section) {
 		cat.HelpLong = helpLong
 	}
 }